+++ Core/Net/AutoUpdate.cs
@@ -22,6 +22,16 @@ namespace CKAN
 
         public static Version FetchLatestCkanVersion()
         {
+            try
+            {
+                FetchUpdaterUrl();
+                FetchCkanUrl();
+            }
+            catch (Kraken)
+            {
+                return new Version(Meta.Version());
+            }
+
             var response = MakeRequest(latestCKANReleaseApiUrl);
 
             return new CKANVersion(response.tag_name.ToString(), response.name.ToString());
@@ -80,16 +90,32 @@ namespace CKAN
             Environment.Exit(0);
         }
 
-        private static Uri FetchUpdaterUrl()
+        // internal so our test suite can poke it.
+        internal static Uri FetchUpdaterUrl(Uri url = null)
         {
-            var response = MakeRequest(latestUpdaterReleaseApiUrl);
+            url = url ?? latestUpdaterReleaseApiUrl;
+
+            var response = MakeRequest(url);
+            if (response.assets.Count == 0)
+            {
+                throw new Kraken("The latest updater isn't uploaded yet.");
+            }
             var assets = response.assets[0];
             return new Uri(assets.browser_download_url.ToString());
         }
 
-        private static Uri FetchCkanUrl()
+        // internal so our test suite can poke it.
+        // TODO: Combine this with FetchUpdaterUrl above, they're
+        // pretty much identical.
+        internal static Uri FetchCkanUrl(Uri url = null)
         {
-            var response = MakeRequest(latestCKANReleaseApiUrl);
+            url = url ?? latestCKANReleaseApiUrl;
+
+            var response = MakeRequest(url);
+            if (response.assets.Count == 0)
+            {
+                throw new Kraken("The latest release isn't uploaded yet.");
+            }
             var assets = response.assets[0];
             return new Uri(assets.browser_download_url.ToString());
         }